モデルクラスのメンバーを Lazy にした
Map から動的に取り出すのでメンバーの取り出し時に例外が発生する可能性があります。
fixes #2

+++ app/models/GeoInfo.scala
@@ -7,7 +7,7 @@ import play.api.libs.json._
 /**
  * Hold geographic location, latitude and longitude, in degrees.
  */
-case class GeoInfo(latitude: BigDecimal, longitude: BigDecimal) {
+case class GeoInfo(latitude: Double, longitude: Double) {
   import GeoInfo._
   object radian {
     lazy val lat = latitude.toDouble.toRadians
